requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.0
google-generativeai>=0.7.0
playwright>=1.53.0
python-dotenv>=1.0.0
